+++ aderyn_driver/src/lib.rs
@@ -4,10 +4,11 @@ pub mod compile;
 pub mod config;
 pub mod driver;
 pub mod lsp_report;
+pub mod preprocess;
 
 use std::path::{Path, PathBuf};
 
-pub use aderyn_core::{ast as core_ast, context, detect as detection_modules, detect::detector};
+pub use aderyn_core::{ast as core_ast, detect as detection_modules, detect::detector};
 
 fn ensure_valid_root_path(root_path: &Path) -> PathBuf {
     if !root_path.exists() {
